refactor: 레이어, 패키지 명 변경
