+++ backend/src/main/java/mouda/backend/notification/exception/NotificationErrorMessage.java
@@ -9,8 +9,7 @@ public enum NotificationErrorMessage {
 
 	// Firebase
 	FCM_TOKEN_NOT_FOUND_BY_MEMBER("회원의 FCM 토큰이 존재하지 않아요."),
-	FCM_TOKEN_NOT_FOUND_BY_TOKEN("요청한 FCM 토큰이 존재하지 않아요.")
-	;
+	FCM_TOKEN_NOT_FOUND_BY_TOKEN("요청한 FCM 토큰이 존재하지 않아요.");
 
 	// Notification
