+++ backend/src/main/java/mouda/backend/notification/implement/notificationbuilder/MoimCancelledNotificationBuilder.java
@@ -1,4 +1,4 @@
-package mouda.backend.notification.domain.notification;
+package mouda.backend.notification.implement.notificationbuilder;
 
 import org.springframework.stereotype.Component;
 
@@ -10,19 +10,20 @@ import mouda.backend.notification.domain.NotificationTypeProvider;
 import mouda.backend.notification.infrastructure.MoudaNotificationRepository;
 
 @Component
-@NotificationTypeProvider(NotificationType.NEW_COMMENT)
-public class NewCommentNotificationBuilder extends MoimNotificationBuilder {
+@NotificationTypeProvider(NotificationType.MOIM_CANCELLED)
+public class MoimCancelledNotificationBuilder extends MoimNotificationBuilder {
 
-	public NewCommentNotificationBuilder(
+	public MoimCancelledNotificationBuilder(
 		MoudaNotificationRepository moudaNotificationRepository) {
 		super(moudaNotificationRepository);
 	}
 
 	@Override
-	public MoudaNotification buildNotification(Long darakbangId, Moim moim, DarakbangMember author) {
+	public MoudaNotification buildNotification(Long darakbangId, Moim moim, DarakbangMember sender) {
+		NotificationType notificationType = NotificationType.MOIM_CANCELLED;
 		MoudaNotification notification = MoudaNotification.builder()
-			.type(NotificationType.NEW_COMMENT)
-			.body(NotificationType.NEW_COMMENT.createMessage(author.getNickname()))
+			.type(notificationType)
+			.body(notificationType.createMessage(moim.getTitle()))
 			.targetUrl(baseUrl + String.format(moimUrl, darakbangId, moim.getId()))
 			.build();
