+++ backend/src/main/java/mouda/backend/notification/implement/notificationbuilder/MoimModifiedNotificationBuilder.java
@@ -1,9 +1,7 @@
-package mouda.backend.notification.domain.notification;
+package mouda.backend.notification.implement.notificationbuilder;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import lombok.RequiredArgsConstructor;
 import mouda.backend.darakbangmember.domain.DarakbangMember;
 import mouda.backend.moim.domain.Moim;
 import mouda.backend.notification.domain.MoudaNotification;
@@ -12,25 +10,21 @@ import mouda.backend.notification.domain.NotificationTypeProvider;
 import mouda.backend.notification.infrastructure.MoudaNotificationRepository;
 
 @Component
-@NotificationTypeProvider(NotificationType.NEW_CHAT)
-@RequiredArgsConstructor
-public class NewChatNotificationBuilder implements NotificationBuilderStrategy {
+@NotificationTypeProvider(NotificationType.MOIM_MODIFIED)
+public class MoimModifiedNotificationBuilder extends MoimNotificationBuilder {
 
-	@Value("${url.base}")
-	private String baseUrl;
-
-	@Value("${url.chatroom}")
-	private String chatroomUrl;
-
-	private final MoudaNotificationRepository moudaNotificationRepository;
+	public MoimModifiedNotificationBuilder(
+		MoudaNotificationRepository moudaNotificationRepository) {
+		super(moudaNotificationRepository);
+	}
 
 	@Override
 	public MoudaNotification buildNotification(Long darakbangId, Moim moim, DarakbangMember sender) {
-		NotificationType notificationType = NotificationType.NEW_CHAT;
+		NotificationType notificationType = NotificationType.MOIM_MODIFIED;
 		MoudaNotification notification = MoudaNotification.builder()
 			.type(notificationType)
-			.body(notificationType.createMessage(sender.getNickname()))
-			.targetUrl(baseUrl + String.format(chatroomUrl, darakbangId, moim.getId()))
+			.body(notificationType.createMessage(moim.getTitle()))
+			.targetUrl(baseUrl + String.format(moimUrl, darakbangId, moim.getId()))
 			.build();
 
 		return moudaNotificationRepository.save(notification);
