+++ backend/src/main/java/mouda/backend/notification/implement/notificationbuilder/MoimNotificationBuilder.java
@@ -1,4 +1,4 @@
-package mouda.backend.notification.domain.notification;
+package mouda.backend.notification.implement.notificationbuilder;
 
 import org.springframework.beans.factory.annotation.Value;
