+++ backend/src/main/java/mouda/backend/notification/implement/notificationbuilder/MoimingCompletedNotificationBuilder.java
@@ -1,4 +1,4 @@
-package mouda.backend.notification.domain.notification;
+package mouda.backend.notification.implement.notificationbuilder;
 
 import org.springframework.stereotype.Component;
 
@@ -10,17 +10,16 @@ import mouda.backend.notification.domain.NotificationTypeProvider;
 import mouda.backend.notification.infrastructure.MoudaNotificationRepository;
 
 @Component
-@NotificationTypeProvider(NotificationType.MOIM_CANCELLED)
-public class MoimCancelledNotificationBuilder extends MoimNotificationBuilder {
-
-	public MoimCancelledNotificationBuilder(
+@NotificationTypeProvider(NotificationType.MOIMING_COMPLETED)
+public class MoimingCompletedNotificationBuilder extends MoimNotificationBuilder {
+	public MoimingCompletedNotificationBuilder(
 		MoudaNotificationRepository moudaNotificationRepository) {
 		super(moudaNotificationRepository);
 	}
 
 	@Override
 	public MoudaNotification buildNotification(Long darakbangId, Moim moim, DarakbangMember sender) {
-		NotificationType notificationType = NotificationType.MOIM_CANCELLED;
+		NotificationType notificationType = NotificationType.MOIMING_COMPLETED;
 		MoudaNotification notification = MoudaNotification.builder()
 			.type(notificationType)
 			.body(notificationType.createMessage(moim.getTitle()))
