+++ backend/src/main/java/mouda/backend/notification/implement/recipientresolver/MoimCanclledNotificationRecipientResolver.java
@@ -1,4 +1,4 @@
-package mouda.backend.notification.domain.recipient;
+package mouda.backend.notification.implement.recipientresolver;
 
 import org.springframework.stereotype.Component;
 
@@ -9,10 +9,10 @@ import mouda.backend.notification.domain.NotificationTypeProvider;
 import mouda.backend.notification.infrastructure.MemberNotificationRepository;
 
 @Component
-@NotificationTypeProvider(NotificationType.MOINING_REOPENED)
-public class MoimingReopenedNotificationRecipientResolver extends MoimStatusChangedNotificationRecipientResolver {
+@NotificationTypeProvider(NotificationType.MOIM_CANCELLED)
+public class MoimCanclledNotificationRecipientResolver extends MoimStatusChangedNotificationRecipientResolver {
 
-	public MoimingReopenedNotificationRecipientResolver(
+	public MoimCanclledNotificationRecipientResolver(
 		DarakbangMemberRepository darakbangMemberRepository,
 		MemberNotificationRepository memberNotificationRepository,
 		ChamyoRepository chamyoRepository) {
