+++ backend/src/main/java/mouda/backend/notification/implement/recipientresolver/MoimStatusChangedNotificationRecipientResolver.java
@@ -1,31 +1,34 @@
-package mouda.backend.notification.domain.recipient;
+package mouda.backend.notification.implement.recipientresolver;
 
 import java.util.List;
-
-import org.springframework.stereotype.Component;
 
 import mouda.backend.darakbangmember.domain.DarakbangMember;
 import mouda.backend.darakbangmember.infrastructure.DarakbangMemberRepository;
 import mouda.backend.moim.domain.Moim;
+import mouda.backend.moim.domain.MoimRole;
 import mouda.backend.moim.infrastructure.ChamyoRepository;
 import mouda.backend.notification.domain.MoudaNotification;
-import mouda.backend.notification.domain.NotificationType;
-import mouda.backend.notification.domain.NotificationTypeProvider;
 import mouda.backend.notification.infrastructure.MemberNotificationRepository;
 
-@Component
-@NotificationTypeProvider(NotificationType.NEW_COMMENT)
-public class NewCommentNotificationRecipientResolver extends NoneChatRecipientResolverStrategy {
-	public NewCommentNotificationRecipientResolver(
+public abstract class MoimStatusChangedNotificationRecipientResolver extends NoneChatRecipientResolverStrategy {
+
+	public MoimStatusChangedNotificationRecipientResolver(
 		DarakbangMemberRepository darakbangMemberRepository,
 		MemberNotificationRepository memberNotificationRepository,
-		ChamyoRepository chamyoRepository) {
+		ChamyoRepository chamyoRepository
+	) {
 		super(darakbangMemberRepository, memberNotificationRepository, chamyoRepository);
 	}
 
 	@Override
 	public List<Long> resolveRecipients(long darakbangId, MoudaNotification notification, Moim moim,
 		DarakbangMember sender) {
-		return List.of(chamyoRepository.findMoimerIdByMoimId(moim.getId()));
+		List<Long> recipientIds = chamyoRepository.findAllByMoimId(moim.getId()).stream()
+			.filter(chamyo -> chamyo.getMoimRole() != MoimRole.MOIMER)
+			.map(chamyo -> chamyo.getDarakbangMember().getMemberId())
+			.toList();
+
+		saveNotificationsForMembers(recipientIds, darakbangId, notification);
+		return recipientIds;
 	}
 }
