+++ backend/src/main/java/mouda/backend/notification/implement/recipientresolver/MoimTimeConfirmedNotificationRecipientResolver.java
@@ -1,4 +1,4 @@
-package mouda.backend.notification.domain.recipient;
+package mouda.backend.notification.implement.recipientresolver;
 
 import java.util.List;
 
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Component;
 import mouda.backend.darakbangmember.domain.DarakbangMember;
 import mouda.backend.darakbangmember.infrastructure.DarakbangMemberRepository;
 import mouda.backend.moim.domain.Moim;
+import mouda.backend.moim.domain.MoimRole;
 import mouda.backend.moim.infrastructure.ChamyoRepository;
 import mouda.backend.notification.domain.MoudaNotification;
 import mouda.backend.notification.domain.NotificationType;
@@ -14,24 +15,22 @@ import mouda.backend.notification.domain.NotificationTypeProvider;
 import mouda.backend.notification.infrastructure.MemberNotificationRepository;
 
 @Component
-@NotificationTypeProvider(NotificationType.NEW_MOIMEE_JOINED)
-public class NewMoimeeJoinedNotificationRecipientResolver extends NoneChatRecipientResolverStrategy {
+@NotificationTypeProvider(NotificationType.MOIM_TIME_CONFIRMED)
+public class MoimTimeConfirmedNotificationRecipientResolver extends NoneChatRecipientResolverStrategy {
 
-	public NewMoimeeJoinedNotificationRecipientResolver(
+	public MoimTimeConfirmedNotificationRecipientResolver(
 		DarakbangMemberRepository darakbangMemberRepository,
 		MemberNotificationRepository memberNotificationRepository,
-		ChamyoRepository chamyoRepository
-	) {
+		ChamyoRepository chamyoRepository) {
 		super(darakbangMemberRepository, memberNotificationRepository, chamyoRepository);
 	}
 
 	@Override
 	public List<Long> resolveRecipients(long darakbangId, MoudaNotification notification, Moim moim,
 		DarakbangMember sender) {
-
 		List<Long> recipientIds = chamyoRepository.findAllByMoimId(moim.getId()).stream()
-			.map(c -> c.getDarakbangMember().getMemberId())
-			.filter(memberId -> !memberId.equals(sender.getMemberId()))
+			.filter(chamyo -> chamyo.getMoimRole() != MoimRole.MOIMER)
+			.map(chamyo -> chamyo.getDarakbangMember().getMemberId())
 			.toList();
 
 		saveNotificationsForMembers(recipientIds, darakbangId, notification);
