+++ backend/src/main/java/mouda/backend/notification/implement/recipientresolver/MoimeeLeftNotificationRecipientResolver.java
@@ -1,4 +1,4 @@
-package mouda.backend.notification.domain.recipient;
+package mouda.backend.notification.implement.recipientresolver;
 
 import java.util.List;
 
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Component;
 import mouda.backend.darakbangmember.domain.DarakbangMember;
 import mouda.backend.darakbangmember.infrastructure.DarakbangMemberRepository;
 import mouda.backend.moim.domain.Moim;
-import mouda.backend.moim.domain.MoimRole;
 import mouda.backend.moim.infrastructure.ChamyoRepository;
 import mouda.backend.notification.domain.MoudaNotification;
 import mouda.backend.notification.domain.NotificationType;
@@ -15,9 +14,10 @@ import mouda.backend.notification.domain.NotificationTypeProvider;
 import mouda.backend.notification.infrastructure.MemberNotificationRepository;
 
 @Component
-@NotificationTypeProvider(NotificationType.MOIM_PLACE_CONFIRMED)
-public class MoimPlaceConfirmedNotificationRecipientResolver extends NoneChatRecipientResolverStrategy {
-	public MoimPlaceConfirmedNotificationRecipientResolver(
+@NotificationTypeProvider(NotificationType.MOIMEE_LEFT)
+public class MoimeeLeftNotificationRecipientResolver extends NoneChatRecipientResolverStrategy {
+
+	public MoimeeLeftNotificationRecipientResolver(
 		DarakbangMemberRepository darakbangMemberRepository,
 		MemberNotificationRepository memberNotificationRepository,
 		ChamyoRepository chamyoRepository) {
@@ -27,12 +27,6 @@ public class MoimPlaceConfirmedNotificationRecipientResolver extends NoneChatRec
 	@Override
 	public List<Long> resolveRecipients(long darakbangId, MoudaNotification notification, Moim moim,
 		DarakbangMember sender) {
-		List<Long> recipientIds = chamyoRepository.findAllByMoimId(moim.getId()).stream()
-			.filter(chamyo -> chamyo.getMoimRole() != MoimRole.MOIMER)
-			.map(chamyo -> chamyo.getDarakbangMember().getMemberId())
-			.toList();
-
-		saveNotificationsForMembers(recipientIds, darakbangId, notification);
-		return recipientIds;
+		return List.of(chamyoRepository.findMoimerIdByMoimId(moim.getId()));
 	}
 }
