+++ backend/src/main/java/mouda/backend/notification/implement/recipientresolver/RecipientResolverStrategy.java
@@ -1,4 +1,4 @@
-package mouda.backend.notification.domain.recipient;
+package mouda.backend.notification.implement.recipientresolver;
 
 import java.util.List;
